Moved the Request parsing functions in the RequestParsing.cpp files.
It will allow me to add more functions to make the parsing better.
