+++ code/srcs/dataFlow/Request.cpp
@@ -1,4 +1,3 @@
-#include <algorithm>              // for find
 #include <cstring>                // for memcmp, size_t, NULL
 #include <exception>              // for exception
 #include <iostream>               // for basic_ostream, operator<<, endl, cout
@@ -71,59 +70,6 @@ int	Request::setBodyFromHeaders(SharedResource<int> destFd, const ServerConfigur
 	}
 	else if (_statusLine.method == PUT || _statusLine.method == POST)
 		return (HTTP_LENGTH_REQUIRED);
-	return (HTTP_OK);
-}
-
-int		Request::parseStatusLine(const char *line, const char *end)
-{
-	//Parsing the method
-	const char	*methodEnd = std::find(line, end, ' ');
-	if (methodEnd == end)
-		return (HTTP_BAD_REQUEST);
-	try {
-		_statusLine.method = getMethodFromBuffer(line, std::distance(line, methodEnd));
-	}
-	catch (std::exception& exception) {
-		return (HTTP_NOT_IMPLEMENTED);
-	}
-
-	//Parsing the target
-	const char	*targetEnd = std::find(methodEnd + 1, end, ' ');
-	if (targetEnd == end)
-		return (HTTP_BAD_REQUEST);
-	_statusLine.requestTarget = std::string(methodEnd + 1, targetEnd);
-
-	//Parsing the protocol
-	const char	*protocolEnd = std::find(targetEnd + 1, end, '\r');
-	if (protocolEnd == end)
-		return (HTTP_BAD_REQUEST);
-	if (std::memcmp(targetEnd + 1, PROTOCOL, PROTOCOL_LENGTH))
-		return (HTTP_HTTP_VERSION_NOT_SUPPORTED);
-	return (HTTP_OK);
-}
-
-static char toLowerCase(char& c)
-{
-    return (std::tolower(c));
-}
-
-int		Request::parseHeader(const char *line, const char *end)
-{
-	if (std::distance(line, end) < 5)
-		return (HTTP_BAD_REQUEST);
-	const char * const keyEnd = std::find(line, end, ':');
-
-	if (keyEnd == end || *(keyEnd + 1) != ' ')
-		return (HTTP_BAD_REQUEST);
-
-	const char * const valueEnd = end - 1;
-	if (*valueEnd != '\r')
-		return (HTTP_BAD_REQUEST);
-	const std::string key(line, keyEnd);
-	const char * valuePosition = keyEnd + 2;
-	std::string value(valuePosition, valueEnd);
-	std::transform(value.begin(), value.end(), value.begin(), toLowerCase);
-	this->_headers[key] = value;
 	return (HTTP_OK);
 }
